refactor(plugin): [account] fix ts errors
