+++ plugins/account/src/components/AccountForm/AccountForm.stories.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
 import { AccountForm } from './AccountForm';
 
 export default {
   title: 'Modules|account/AccountForm',
 };
 
-export const accountForm = () => <AccountForm currentAccount={{}} />;
+export const accountForm = () => <AccountForm handleSubmit={() => {}} currentAccount={{}} />;
