+++ plugins/account/src/components/AccountSelector.tsx
@@ -12,13 +12,16 @@ export const AccountSelector: React.FC = () => {
   const currentUser = useSelector((store: CoreStore) => store.authStore.currentUser);
 
   useEffect(() => {
+    //@ts-ignore
     if (currentUser.accounts) {
+      //@ts-ignore
       setAccounts(currentUser.accounts);
 
       if (accounts[0] && !accountId) {
         changeAccount(accounts[0].id);
       }
     }
+    // @ts-ignore
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [currentUser.accounts]);
 
@@ -37,7 +40,7 @@ export const AccountSelector: React.FC = () => {
         onChange={changeAccount}
         value={accountId}
         filterOption={(input, option) =>
-          option.props.children.toLowerCase().indexOf(input.toLowerCase()) >= 0
+          option?.props.children.toLowerCase().indexOf(input.toLowerCase()) >= 0
         }
       >
         {accounts.map(account => (
