--- plugins/account/src/documents/AccountApiKey.mdx
@@ -1,12 +1,0 @@
-import { Descriptions } from 'antd';
-
-## API Authentication
-
-<Descriptions>
-  <Descriptions.Item label="Header Key" span={3}>
-    Authorization
-  </Descriptions.Item>
-  <Descriptions.Item label="Header Value" span={3}>
-    Bearer {props.accountApiKey.token}
-  </Descriptions.Item>
-</Descriptions>
