--- plugins/account/src/documents/index.ts
@@ -1,5 +1,0 @@
-import AccountApiKey from './AccountApiKey.mdx';
-
-export * from './Authentication';
-
-export { AccountApiKey };
