+++ plugins/account/src/entities/interfaces/IAccount.ts
