+++ plugins/account/src/entities/interfaces/index.ts
@@ -1,1 +1,1 @@
-export * from './IAccount.d';
+export * from './IAccount';
