+++ plugins/account/src/index.ts
@@ -1,5 +1,6 @@
 export * from './components';
 export * from './entities';
 export * from './pages';
-export * from './services';
 export * from './redux';
+export * from './services';
+export * from './__mocks__';
