+++ plugins/account/src/pages/AccountListPage.tsx
@@ -2,6 +2,7 @@ import { CoreStore, DefaultPubSubContext } from '@onr/core';
 import React, { useContext, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { AccountList } from '../components/AccountList';
+import { accountActions } from '../redux';
 import { StoreProps } from '../redux/types';
 
 type Store = CoreStore & StoreProps;
@@ -9,6 +10,7 @@ type Store = CoreStore & StoreProps;
 export const AccountListPage: React.FC = () => {
   const dispatch = useDispatch();
   const accounts = useSelector((store: Store) => store.accountStore.accounts);
+  // @ts-ignore
   const { subscribe } = useContext(DefaultPubSubContext);
 
   useEffect(() => {
