+++ plugins/account/src/services/interfaces/AccountModel.ts
@@ -1,0 +1,11 @@
+import { AccountsPayload, AccountsResponse } from '.';
+import { IAccount } from '../../entities/interfaces';
+
+export interface GetAccountsPayload {
+  params: AccountsPayload;
+}
+
+export interface GetAccountsResponse extends AccountsResponse {}
+export interface GetAccountResponse {
+  data: IAccount;
+}
