+++ plugins/account/src/services/interfaces/AccountsResponse.ts
@@ -10,5 +10,5 @@ export interface AccountsResponse {
    * @type {array|null}
    * @memberof AccountsResponse
    */
-  data: array;
+  data: [];
 }
